+++ library/src/main/java/com/twilio/video/VideoCodec.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) 2017 Twilio, inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.twilio.video;
+
+/**
+ * Definition of supported video codecs.
+ */
+public enum VideoCodec {
+    /**
+     * Block-oriented motion-compensation-based video compression standard.
+     *
+     * @see <a href="https://en.wikipedia.org/wiki/H.264/MPEG-4_AVC">H.264</a>
+     */
+    H264,
+
+    /**
+     * Traditional block-based transform coding format similar to H264.
+     *
+     * @see <a href="https://en.wikipedia.org/wiki/VP8">VP8</a>
+     */
+    VP8,
+
+    /**
+     * Traditional block-based transform coding format similar to MPEG's High Efficiency Video
+     * Coding (HEVC/H.265).
+     *
+     * @see <a href="https://en.wikipedia.org/wiki/VP9">VP9</a>
+     */
+    VP9
+}