+++ library/src/main/jni/com_twilio_video_ConnectOptions.cpp
@@ -25,6 +25,58 @@
 
 namespace twilio_video_jni {
 
+twilio::media::AudioCodec getAudioCodec(JNIEnv *env, jobject j_audio_codec) {
+    jclass j_audio_codec_class = webrtc_jni::GetObjectClass(env, j_audio_codec);
+    jmethodID j_name_method_id = webrtc_jni::GetMethodID(env,
+                                                         j_audio_codec_class,
+                                                         "name",
+                                                         "()Ljava/lang/String;");
+    jstring j_audio_codec_name = (jstring) env->CallObjectMethod(j_audio_codec, j_name_method_id);
+    CHECK_EXCEPTION(env) << "Failed to get name of audio codec";
+    std::string audio_codec_name = webrtc_jni::JavaToStdString(env, j_audio_codec_name);
+    twilio::media::AudioCodec audio_codec;
+
+    if (audio_codec_name == "ISAC") {
+        audio_codec = twilio::media::AudioCodec::ISAC;
+    } else if (audio_codec_name == "OPUS") {
+        audio_codec = twilio::media::AudioCodec::OPUS;
+    } else if (audio_codec_name == "PCMA") {
+        audio_codec = twilio::media::AudioCodec::PCMA;
+    } else if (audio_codec_name == "PCMU") {
+        audio_codec = twilio::media::AudioCodec::PCMU;
+    } else if (audio_codec_name == "G722") {
+        audio_codec = twilio::media::AudioCodec::G722;
+    } else {
+        FATAL() << "Failed to get native audio codec for " << audio_codec_name;
+    }
+
+    return audio_codec;
+}
+
+twilio::media::VideoCodec getVideoCodec(JNIEnv *env, jobject j_video_codec) {
+    jclass j_video_codec_class = webrtc_jni::GetObjectClass(env, j_video_codec);
+    jmethodID j_name_method_id = webrtc_jni::GetMethodID(env,
+                                                         j_video_codec_class,
+                                                         "name",
+                                                         "()Ljava/lang/String;");
+    jstring j_video_codec_name = (jstring) env->CallObjectMethod(j_video_codec, j_name_method_id);
+    CHECK_EXCEPTION(env) << "Failed to get name of video codec";
+    std::string video_codec_name = webrtc_jni::JavaToStdString(env, j_video_codec_name);
+    twilio::media::VideoCodec video_codec;
+
+    if (video_codec_name == "H264") {
+        video_codec = twilio::media::VideoCodec::H264;
+    } else if (video_codec_name == "VP8") {
+        video_codec = twilio::media::VideoCodec::VP8;
+    } else if (video_codec_name == "VP9") {
+        video_codec = twilio::media::VideoCodec::VP9;
+    } else {
+        FATAL() << "Failed to get native video codec for " << video_codec_name;
+    }
+
+    return video_codec;
+}
+
 JNIEXPORT jlong JNICALL
 Java_com_twilio_video_ConnectOptions_nativeCreate(JNIEnv *env,
                                                   jobject j_instance,
@@ -34,7 +86,9 @@ Java_com_twilio_video_ConnectOptions_nativeCreate(JNIEnv *env,
                                                   jobjectArray j_video_tracks,
                                                   jobject j_ice_options,
                                                   jboolean j_enable_insights,
-                                                  jlong j_platform_info_handle) {
+                                                  jlong j_platform_info_handle,
+                                                  jobjectArray j_preferred_audio_codecs,
+                                                  jobjectArray j_preferred_video_codecs) {
 
     std::string access_token = webrtc_jni::JavaToStdString(env, j_access_token);
     twilio::video::ConnectOptions::Builder* builder =
@@ -88,6 +142,34 @@ Java_com_twilio_video_ConnectOptions_nativeCreate(JNIEnv *env,
                 video_tracks.push_back(video_track);
             }
             builder->setVideoTracks(video_tracks);
+        }
+    }
+
+    if (!webrtc_jni::IsNull(env, j_preferred_audio_codecs)) {
+        std::vector<twilio::media::AudioCodec> preferred_audio_codecs;
+        int size = env->GetArrayLength(j_preferred_audio_codecs);
+        if (size > 0) {
+            for (int i = 0; i < size; i++) {
+                jobject j_audio_codec =
+                        (jobject) env->GetObjectArrayElement(j_preferred_audio_codecs, i);
+                twilio::media::AudioCodec audio_codec = getAudioCodec(env, j_audio_codec);
+                preferred_audio_codecs.push_back(audio_codec);
+            }
+            builder->setPreferredAudioCodecs(preferred_audio_codecs);
+        }
+    }
+
+    if (!webrtc_jni::IsNull(env, j_preferred_video_codecs)) {
+        std::vector<twilio::media::VideoCodec> preferred_video_codecs;
+        int size = env->GetArrayLength(j_preferred_video_codecs);
+        if (size > 0) {
+            for (int i = 0; i < size; i++) {
+                jobject j_video_codec =
+                        (jobject) env->GetObjectArrayElement(j_preferred_video_codecs, i);
+                twilio::media::VideoCodec video_codec = getVideoCodec(env, j_video_codec);
+                preferred_video_codecs.push_back(video_codec);
+            }
+            builder->setPreferredVideoCodecs(preferred_video_codecs);
         }
     }